Langchain_Langgraph_Project/5-langgraph-agent/4langsmithInvoiceAgent/requirements.txt
langchain
python-dotenv
langchain-openai
langchain-core
langchain-community
bs4
faiss-cpu
pypdf
arxiv
pymupdf
wikipedia
lxml
langchain_huggingface
langchain-groq
langgraph
langgraph-cli[inmem]
fastapi
uvicorn[standard]
jupyter
ipykernel
chromadb
matplotlib
ipython
mermaid-py